feat: added broker_funcs param on UploaderBuilder, removed trigger_broker clutter
